Progress output
Version .2 introduces a progress output indicator to better visualize progress.
Progress is indicated by what file is currently being worked on and periodically will say what line is being worked on.
Note:
This will reduce the overall speed because in order to accurately determine (at-least from my current research) how many lines are in each file is to actually check it. Another way could be to simply use a buffered reader so I know how much data has been read and then can just compare that to the file size. I may change it but honestly sequentially reading a file is pretty quick. In the future I plan to add the ability to make this an option or more likely just output what file is being worked on.